Move QgsElevationProfileLayerTreeView to GUI library
