+++ src/app/CMakeLists.txt
@@ -180,8 +180,6 @@ set(QGIS_APP_SRCS
 
   elevation/qgselevationprofileexportsettingswidget.cpp
   elevation/qgselevationprofileimageexportdialog.cpp
-  elevation/qgselevationprofilelayertreemodel.cpp
-  elevation/qgselevationprofilelayertreeview.cpp
   elevation/qgselevationprofilepdfexportdialog.cpp
   elevation/qgselevationprofiletoolidentify.cpp
   elevation/qgselevationprofiletoolmeasure.cpp
